services/web/project/requirements.txt
Quart==0.8.1
hypercorn==0.5.3
redis==2.10.6
websocket-client==0.54.0
python-dotenv==0.10.1
bittrex-websocket==1.0.6.3
python-binance==0.7.1
kafka-python==1.4.4
faust[eventlet]==1.4.6
eventlet==0.24.1